requirements.txt
python-telegram-bot==13.15
Flask==2.3.3
python-dotenv==1.0.1
requests
